+++ tests/bootstrap.php
@@ -2,5 +2,5 @@
 
 require_once __DIR__ . '/../../../tests/bootstrap.php';
 
-\OC_App::loadApp(OCA\Github\AppInfo\Application::APP_ID);
+\OC_App::loadApps();
 OC_Hook::clear();